+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Change Log
 
+## [1.1.0]
+### Added
+- Hamster hordes!!!
+- Randomised durations
+
 ## [1.0.0]
-### [Added]
+### Added
 - First release
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)